+++ src/main/java/com/braintreegateway/Transaction.java
@@ -180,6 +180,8 @@ public class Transaction {
     private String authorizedTransactionId;
     private List<String> partialSettlementTransactionIds;
     private List<AuthorizationAdjustment> authorizationAdjustments;
+    private FacilitatedDetails facilitatedDetails;
+    private FacilitatorDetails facilitatorDetails;
 
     public Transaction(NodeWrapper node) {
         amount = node.findBigDecimal("amount");
@@ -312,6 +314,16 @@ public class Transaction {
         authorizationAdjustments = new ArrayList<AuthorizationAdjustment>();
         for (NodeWrapper authorizationAdjustmentNode : node.findAll("authorization-adjustments/authorization-adjustment")) {
             authorizationAdjustments.add(new AuthorizationAdjustment(authorizationAdjustmentNode));
+        }
+
+        NodeWrapper facilitatedDetailsNode = node.findFirst("facilitated-details");
+        if (facilitatedDetailsNode != null) {
+            facilitatedDetails = new FacilitatedDetails(facilitatedDetailsNode);
+        }
+
+        NodeWrapper facilitatorDetailsNode = node.findFirst("facilitator-details");
+        if (facilitatorDetailsNode != null) {
+            facilitatorDetails = new FacilitatorDetails(facilitatorDetailsNode);
         }
     }
 
@@ -593,5 +605,13 @@ public class Transaction {
 
     public List<AuthorizationAdjustment> getAuthorizationAdjustments() {
         return authorizationAdjustments;
+    }
+
+    public FacilitatedDetails getFacilitatedDetails() {
+        return facilitatedDetails;
+    }
+
+    public FacilitatorDetails getFacilitatorDetails() {
+        return facilitatorDetails;
     }
 }